+++ sqlalchemy_continuum/exc.py
@@ -1,0 +1,6 @@
+class VersioningError(Exception):
+    pass
+
+
+class ClassNotVersioned(VersioningError):
+    pass